Fixed cache listener when filename is empty

+++ src/cuddlybuddly/thumbnail/listeners.py
@@ -10,9 +10,11 @@ def update_cache(sender, instance, **kwargs):
         field = getattr(instance, field)
         if isinstance(field, FieldFile):
             cache_dir = getattr(settings, 'CUDDLYBUDDLY_THUMBNAIL_CACHE', '')
-            cache = os.path.join(
-                cache_dir,
-                md5_constructor(force_unicode(field)).hexdigest()
-            )
-            if os.path.exists(cache):
-                os.remove(cache)
+            field = force_unicode(field)
+            if field:
+                cache = os.path.join(
+                    cache_dir,
+                    md5_constructor(field).hexdigest()
+                )
+                if os.path.exists(cache):
+                    os.remove(cache)
